RDCC-1660: amending functional test to reflect changes
